Converted ubiquo_design to a Rails 3 Engine

+++ ubiquo_design/install/db/migrate/20090224153407_add_plugin_permissions.rb
@@ -1,10 +1,14 @@
 class AddPluginPermissions < ActiveRecord::Migration
   def self.up
-    Permission.create :key => "design_management", :name => "Design management"
-    Permission.create :key => "sitemap_management", :name => "Sitemap management"
+    if const_defined?(:Permission)
+      Permission.create :key => "design_management", :name => "Design management"
+      Permission.create :key => "sitemap_management", :name => "Sitemap management"
+    end
   end
 
   def self.down
-    Permission.destroy_all(:key => %w[design_management sitemap_management])
+    if const_defined?(:Permission)
+      Permission.destroy_all(:key => %w[design_management sitemap_management])
+    end
   end
 end
